Another rag-tag batch of compiler warning removals.
90% of these are unused variables, but there's a few missing
prototypes and the occasional printf arguments issue (in error
handling code).


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2520 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1

+++ src/Misc/dstring.c
@@ -583,8 +583,8 @@ int dstring_htmlise_links(dstring_t *ds) {
 		return -1;
 
 	    if (-1 == dstring_insertf(url, 0, "<a href=\"%.*s\">%.*s</a>",
-				      end-pos, &str[pos],
-				      end-pos, &str[pos])) {
+				      (int)(end-pos), &str[pos],
+				      (int)(end-pos), &str[pos])) {
 		dstring_destroy(url);
 		return -1;
 	    }
